Refactored views into their own package
