+++ psbackend/views/__init__.py
@@ -1,0 +1,4 @@
+from psbackend.views import staticviews
+from psbackend.views import userviews
+
+__all__ = ["staticviews", "userviews"]
